% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_model_metrics.R
\name{compute_model_metrics}
\alias{compute_model_metrics}
\title{Compute evaluation metrics from a confusion matrix}
\usage{
compute_model_metrics(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{A 2x2 matrix with named rows and columns}
}
\value{
A data frame with model evaluation metrics
}
\description{
Compute evaluation metrics from a confusion matrix
}
