% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download a dataset from a URL and save it locally}
\usage{
download_data(url, output)
}
\arguments{
\item{url}{A string specifying the URL of the dataset}

\item{output}{A string specifying the local file path to save the dataset}
}
\value{
No return value, called for side effects (saves file locally)
}
\description{
Download a dataset from a URL and save it locally
}
\examples{
download_data("https://example.com/data.csv", "data/adult.csv")
}
