% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_logistic_model.R
\name{train_logistic_model}
\alias{train_logistic_model}
\title{Train a logistic regression model for binary classification}
\usage{
train_logistic_model(
  data,
  outcome_var,
  predictor_vars,
  pos_class = ">50K",
  neg_class = "<=50K"
)
}
\arguments{
\item{data}{A data frame containing the variables to be used in the model}

\item{outcome_var}{A string specifying the name of the binary outcome variable}

\item{predictor_vars}{A character vector specifying the names of predictor variables}

\item{pos_class}{A string specifying the positive class in the outcome variable}

\item{neg_class}{A string specifying the negative class in the outcome variable}
}
\value{
A list containing the trained model and the ROC curve object
}
\description{
This function trains a logistic regression model on the provided dataset
using the specified predictor variables and binary outcome variable.
}
