% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Split a data frame into training and testing subsets}
\usage{
train_test_split(data, train_fraction = 0.8)
}
\arguments{
\item{data}{A data frame or tibble to be split into training and testing subsets}

\item{train_fraction}{A numeric value between 0 and 1 specifying the fraction of rows for training}
}
\value{
A list containing two elements:
\item{train}{The training subset as a data frame}
\item{test}{The testing subset as a data frame}
}
\description{
Split a data frame into training and testing subsets
}
\examples{
# Given an existing data frame called df
# result <- train_test_split(df, 0.8)
# head(result$train)
# head(result$test)
}
